// src/model/MonitoringData.ts
export interface MonitoringData {
  username: string;
  Id: number;
  Market: string;
  Division: string;
  Channel: string;
  Retailer: string;
  UploadDate: Date;
  ProcessDate: Date;
  UploadChannel: string;
  FilesName: string;
  Execution: string;
  FailureReason: string;
  IncidentNo: string;
  Status: string;
  TimeStamp: string;
  showMore: boolean;
}